refactor: rename Poke components and align tool rows
